fix: preserve runtime failure evidence
Changed:
Centralized browser diagnostics, captured authentication failures, recorded dependency failures, and supplied one GitHub-run correlation identifier to every worker.

Benefit:
Runtime Harness failures retain complete, run-scoped evidence across setup and smoke phases.
