+++ src/cook-web/e2e/smoke.spec.ts
@@ -1,44 +1,13 @@
-import { expect, Page, TestInfo, test } from '@playwright/test';
-import { mkdir, writeFile } from 'node:fs/promises';
+import { expect, Page, test } from '@playwright/test';
 
-type ConsoleEntry = {
-  type: string;
-  text: string;
-};
-
-type NetworkEntry = {
-  method: string;
-  resourceType: string;
-  status: number | null;
-  url: string;
-  requestId: string;
-  harnessRunId: string;
-};
+import {
+  attachRuntimeHarnessDiagnostics,
+  type NetworkEntry,
+  type RuntimeHarnessDiagnostics,
+} from './harness-diagnostics';
 
 const DEFAULT_DEMO_SHIFU_BID =
   process.env.AI_SHIFU_DEMO_SHIFU_BID || 'b5d7844387e940ed9480a6f945a6db6a';
-const HARNESS_RUN_ID =
-  process.env.AI_SHIFU_HARNESS_RUN_ID || `pw-run-${Date.now()}`;
-
-const createRequestId = (testInfo: TestInfo) =>
-  `pw-${Date.now()}-${testInfo.title
-    .toLowerCase()
-    .replace(/[^a-z0-9]+/g, '-')
-    .replace(/^-|-$/g, '')
-    .slice(0, 32)}`;
-
-const buildRuntimeHarnessHints = (
-  requestId: string,
-  harnessRunId: string,
-  diagnosticsPath: string,
-) => ({
-  requestId,
-  harnessRunId,
-  browserDiagnostics: diagnosticsPath,
-  composeLog: 'artifacts/runtime-harness/compose.log',
-  composeStatus: 'artifacts/runtime-harness/compose-ps.json',
-  investigation: 'Review the uploaded runtime-harness-artifacts bundle.',
-});
 
 const waitForCourseData = async (page: Page, entries: NetworkEntry[]) => {
   if (
@@ -83,108 +52,14 @@ const expectNoServerErrors = (serverErrorEntries: NetworkEntry[]) => {
 };
 
 test.describe('agent-first smoke harness', () => {
-  let consoleEntries: ConsoleEntry[] = [];
-  let networkEntries: NetworkEntry[] = [];
-  let serverErrorEntries: NetworkEntry[] = [];
-  let lastObservedRequestId = '';
+  let diagnostics: RuntimeHarnessDiagnostics;
 
   test.beforeEach(async ({ page }, testInfo) => {
-    consoleEntries = [];
-    networkEntries = [];
-    serverErrorEntries = [];
-
-    const requestId = createRequestId(testInfo);
-    lastObservedRequestId = requestId;
-    await page.context().setExtraHTTPHeaders({
-      'X-Request-ID': requestId,
-      'X-Harness-Run-ID': HARNESS_RUN_ID,
-    });
-    await page.addInitScript(harnessRunId => {
-      (window as any).__HARNESS_RUN_ID__ = harnessRunId;
-      window.sessionStorage.setItem('harness_run_id', String(harnessRunId));
-    }, HARNESS_RUN_ID);
-
-    page.on('console', message => {
-      consoleEntries.push({
-        type: message.type(),
-        text: message.text(),
-      });
-      if (consoleEntries.length > 40) {
-        consoleEntries = consoleEntries.slice(-40);
-      }
-    });
-
-    page.on('response', response => {
-      const request = response.request();
-      const headers = request.headers();
-      const requestIdHeader = headers['x-request-id'];
-      if (requestIdHeader) {
-        lastObservedRequestId = requestIdHeader;
-      }
-      const entry = {
-        method: request.method(),
-        resourceType: request.resourceType(),
-        status: response.status(),
-        url: response.url(),
-        requestId: requestIdHeader || lastObservedRequestId,
-        harnessRunId: headers['x-harness-run-id'] || HARNESS_RUN_ID,
-      };
-      networkEntries.push(entry);
-      if (entry.url.includes('/api/') && entry.status >= 500) {
-        serverErrorEntries.push(entry);
-      }
-      if (networkEntries.length > 60) {
-        networkEntries = networkEntries.slice(-60);
-      }
-    });
-
-    page.on('requestfailed', request => {
-      networkEntries.push({
-        method: request.method(),
-        resourceType: request.resourceType(),
-        status: null,
-        url: request.url(),
-        requestId: lastObservedRequestId,
-        harnessRunId: HARNESS_RUN_ID,
-      });
-      if (networkEntries.length > 60) {
-        networkEntries = networkEntries.slice(-60);
-      }
-    });
+    diagnostics = await attachRuntimeHarnessDiagnostics(page, testInfo);
   });
 
-  test.afterEach(async ({ page }, testInfo) => {
-    if (testInfo.status === testInfo.expectedStatus) {
-      return;
-    }
-
-    await mkdir(testInfo.outputDir, { recursive: true });
-
-    const screenshotPath = testInfo.outputPath('failure.png');
-    await page.screenshot({ path: screenshotPath, fullPage: true });
-
-    const diagnosticsPath = testInfo.outputPath('harness-diagnostics.json');
-    await writeFile(
-      diagnosticsPath,
-      JSON.stringify(
-        {
-          pageUrl: page.url(),
-          harnessRunId: HARNESS_RUN_ID,
-          lastRequestId: lastObservedRequestId,
-          console: consoleEntries,
-          network: networkEntries.slice(-25),
-          screenshot: screenshotPath,
-          runtimeHarness: buildRuntimeHarnessHints(
-            lastObservedRequestId,
-            HARNESS_RUN_ID,
-            diagnosticsPath,
-          ),
-        },
-        null,
-        2,
-      ),
-      'utf-8',
-    );
+  test.afterEach(async ({}, testInfo) => {
+    await diagnostics.captureFailure(testInfo);
   });
 
   test('authenticated admin operations page loads without server errors', async ({
@@ -196,7 +71,7 @@ test.describe('agent-first smoke harness', () => {
     await expect(page.getByTestId('admin-operations-header')).toBeVisible();
     await expect(page.getByTestId('admin-operations-filters')).toBeVisible();
     await adminBootstrap;
-    expectNoServerErrors(serverErrorEntries);
+    expectNoServerErrors(diagnostics.serverErrorEntries);
   });
 
   test('learner course shell loads its first data request without server errors', async ({
@@ -205,7 +80,7 @@ test.describe('agent-first smoke harness', () => {
     const coursePath = `/c/${DEFAULT_DEMO_SHIFU_BID}`;
     await page.goto(coursePath);
     await expect(page.getByTestId('course-chat-page')).toBeVisible();
-    await waitForCourseData(page, networkEntries);
-    expectNoServerErrors(serverErrorEntries);
+    await waitForCourseData(page, diagnostics.networkEntries);
+    expectNoServerErrors(diagnostics.serverErrorEntries);
   });
 });
